Refactor: DTO 변환 메서드를 서비스 계층으로 이동
