+++ backend/src/main/java/kr/codesquad/sidedish/service/CategoryService.java
@@ -1,14 +1,13 @@
 package kr.codesquad.sidedish.service;
 
+import kr.codesquad.sidedish.controller.dto.CategoryListResponse;
 import kr.codesquad.sidedish.domain.Category;
 import kr.codesquad.sidedish.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
 @Service
 @RequiredArgsConstructor
@@ -16,7 +15,22 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
 
-    public List<Category> readAll() {
-        return categoryRepository.findAll();
+    public List<CategoryListResponse.Element> readAll() {
+        return categoryRepository.findAll()
+                .stream()
+                .map(this::entityToDTO)
+                .collect(Collectors.toList());
+    }
+
+    private CategoryListResponse.Element entityToDTO(Category category) {
+        List<CategoryListResponse.SubElement> subElements = category.getSubcategories().stream()
+            .map(subcategory -> new CategoryListResponse.SubElement(subcategory.getId(), subcategory.getName()))
+            .collect(Collectors.toList());
+
+        return new CategoryListResponse.Element(
+            category.getId(),
+            category.getName(),
+            category.getDescription(),
+            subElements);
     }
 }
